refactor: mvc 패턴 적용에 따른 폴더 구조 재정리
- models폴더 생성
- controller 폴더 생성
